refactor(checkout): colocate payment components
