+++ src/app/(main)/(checkout)/_components/index.tsx
@@ -1,1 +1,2 @@
 export * from "./OrderSummary";
+export * from "./PageTitle";
