+++ src/app/(main)/(checkout)/payment/_components/index.tsx
@@ -3,4 +3,3 @@ export { ShippingInfoCard } from "./ShippingInfoCard";
 export { CheckoutSubmitBar } from "./CheckoutSubmitBar";
 export { TermsAgreementCard } from "./TermsAgreementCard";
 export { PaymentPendingOverlay } from "./PaymentPendingOverlay";
-export { CheckoutForm } from "./CheckoutForm";
