+++ src/app/(main)/(checkout)/payment/_containers/index.tsx
@@ -1,0 +1,1 @@
+export { CheckoutForm } from "./CheckoutForm";
