+++ src/ui/components/organisms/index.ts
@@ -3,7 +3,6 @@ export * from "./BasicInfoForm";
 export * from "./BasicInfoSection";
 export * from "./BottomActionBar";
 export * from "./ChangePasswordForm";
-export * from "./CheckoutForm";
 export * from "./ClipboardButton";
 export * from "./ComingSoonTemplate";
 export * from "./InvitationFormView";
@@ -17,7 +16,6 @@ export * from "./ImagePreviewItem";
 export * from "./ImagesSection";
 export * from "./ParentsInfoSection";
 export * from "./PersonValueCard";
-export * from "./PaymentMethodSelector";
 export * from "./PremiumFeatureDialog";
 export * from "./PremiumFeatureRegistrationForm";
 export * from "./ProductCatalog";
